Added a fancy Rofi setup script
runrofi.sh now takes SS of screen, blurs it, and then runs rofi with that as the background. Rofi is now full screen. install.sh was changed to copy the runrofi.sh file to /usr/bin. config was changed so when +L was pressed, runrofi.sh would run instead of the other command used. Running rofi this way takes more time though, so remove the scrot & convert commands and remove the -fake-background option from the rofi command if you don't want it to take the screenshot and all that fancy stuff.